Consolidate sprite defs and generation
