+++ MegaManLofi/MegaManLofi.vcxproj
@@ -151,7 +151,7 @@
   <ItemGroup>
     <ClCompile Include="Arena.cpp" />
     <ClCompile Include="ArenaDefsGenerator.cpp" />
-    <ClCompile Include="ArenaSpriteGenerator.cpp" />
+    <ClCompile Include="ArenaImageGenerator.cpp" />
     <ClCompile Include="ArenaTileGenerator.cpp" />
     <ClCompile Include="ArenaPhysics.cpp" />
     <ClCompile Include="BasicEntity.cpp" />
@@ -159,6 +159,7 @@
     <ClCompile Include="ConsoleBuffer.cpp" />
     <ClCompile Include="ConsoleRenderDefsGenerator.cpp" />
     <ClCompile Include="ConsoleSprite.cpp" />
+    <ClCompile Include="ConsoleSpriteDefsGenerator.cpp" />
     <ClCompile Include="EntityConsoleSprite.cpp" />
     <ClCompile Include="EntityDefsGenerator.cpp" />
     <ClCompile Include="EntityFactory.cpp" />
@@ -179,13 +180,12 @@
     <ClCompile Include="KeyboardInputReader.cpp" />
     <ClCompile Include="KeyboardWrapper.cpp" />
     <ClCompile Include="MenuRepository.cpp" />
-    <ClCompile Include="MenuSpriteGenerator.cpp" />
+    <ClCompile Include="MenuImageGenerator.cpp" />
     <ClCompile Include="PlayerDefsGenerator.cpp" />
     <ClCompile Include="PlayerExplodedConsoleAnimation.cpp" />
     <ClCompile Include="PlayerPhysics.cpp" />
     <ClCompile Include="Player.cpp" />
     <ClCompile Include="PlayerPhysicsDefsGenerator.cpp" />
-    <ClCompile Include="PlayerSpriteGenerator.cpp" />
     <ClCompile Include="PlayerThwipInConsoleAnimation.cpp" />
     <ClCompile Include="PlayerThwipOutConsoleAnimation.cpp" />
     <ClCompile Include="PlayingMenu.cpp" />
@@ -200,14 +200,14 @@
     <ClCompile Include="TitleStateConsoleRenderer.cpp" />
     <ClCompile Include="TitleStateInputHandler.cpp" />
     <ClCompile Include="ThreadWrapper.cpp" />
-    <ClCompile Include="TitleSpriteGenerator.cpp" />
+    <ClCompile Include="TitleImageGenerator.cpp" />
     <ClCompile Include="UniqueNumberGenerator.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Arena.h" />
     <ClInclude Include="ArenaDefs.h" />
     <ClInclude Include="ArenaDefsGenerator.h" />
-    <ClInclude Include="ArenaSpriteGenerator.h" />
+    <ClInclude Include="ArenaImageGenerator.h" />
     <ClInclude Include="ArenaGenerationStrings.h" />
     <ClInclude Include="ArenaTileGenerator.h" />
     <ClInclude Include="ArenaPhysics.h" />
@@ -220,6 +220,8 @@
     <ClInclude Include="ConsoleImage.h" />
     <ClInclude Include="ConsoleRenderDefsGenerator.h" />
     <ClInclude Include="ConsoleSprite.h" />
+    <ClInclude Include="ConsoleSpriteDefs.h" />
+    <ClInclude Include="ConsoleSpriteDefsGenerator.h" />
     <ClInclude Include="Coordinate.h" />
     <ClInclude Include="DiagnosticsConsoleRenderer.h" />
     <ClInclude Include="Direction.h" />
@@ -291,7 +293,7 @@
     <ClInclude Include="KeyboardInputReader.h" />
     <ClInclude Include="KeyCode.h" />
     <ClInclude Include="MenuRepository.h" />
-    <ClInclude Include="MenuSpriteGenerator.h" />
+    <ClInclude Include="MenuImageGenerator.h" />
     <ClInclude Include="MenuType.h" />
     <ClInclude Include="MovementType.h" />
     <ClInclude Include="PlayerDefsGenerator.h" />
@@ -299,7 +301,6 @@
     <ClInclude Include="PlayerPhysics.h" />
     <ClInclude Include="PlayerPhysicsDefs.h" />
     <ClInclude Include="PlayerPhysicsDefsGenerator.h" />
-    <ClInclude Include="PlayerSpriteGenerator.h" />
     <ClInclude Include="PlayerThwipInConsoleAnimation.h" />
     <ClInclude Include="PlayerThwipOutConsoleAnimation.h" />
     <ClInclude Include="PlayingMenu.h" />
@@ -321,7 +322,7 @@
     <ClInclude Include="TitleStateConsoleRenderer.h" />
     <ClInclude Include="ThreadWrapper.h" />
     <ClInclude Include="Quad.h" />
-    <ClInclude Include="TitleSpriteGenerator.h" />
+    <ClInclude Include="TitleImageGenerator.h" />
     <ClInclude Include="UniqueNumberGenerator.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
