+++ MegaManLofi/MenuImageGenerator.h
@@ -9,10 +9,9 @@ namespace MegaManLofi
    class IConsoleSprite;
    class IFrameRateProvider;
 
-   class MenuSpriteGenerator
+   class MenuImageGenerator
    {
    public:
       static ConsoleImage GeneratePlayerImage();
-      static std::shared_ptr<IConsoleSprite> GenerateMenuCaratSprite( const std::shared_ptr<IFrameRateProvider> frameRateProvider );
    };
 }
