+++ MegaManLofiTests/PlayerExplodedConsoleAnimationTests.cpp
@@ -4,6 +4,7 @@
 
 #include <MegaManLofi/PlayerExplodedConsoleAnimation.h>
 #include <MegaManLofi/ConsoleRenderDefs.h>
+#include <MegaManLofi/ConsoleSpriteDefs.h>
 
 #include "mock_ConsoleBuffer.h"
 #include "mock_FrameRateProvider.h"
@@ -20,13 +21,16 @@ public:
    {
       _consoleBufferMock.reset( new NiceMock<mock_ConsoleBuffer> );
       _frameRateProviderMock.reset( new NiceMock<mock_FrameRateProvider> );
+      _spriteDefs.reset( new ConsoleSpriteDefs );
       _renderDefs.reset( new ConsoleRenderDefs );
       _particleSpriteMock.reset( new NiceMock<mock_ConsoleSprite> );
 
-      _renderDefs->PlayerExplosionParticleSprite = _particleSpriteMock;
+      _spriteDefs->PlayerExplosionParticleSprite = _particleSpriteMock;
+
       _renderDefs->PlayerExplosionParticleVelocity = 2;
       _renderDefs->ArenaCharWidth = 1;
       _renderDefs->ArenaCharHeight = 1;
+      _renderDefs->SpriteDefs = _spriteDefs;
 
       ON_CALL( *_frameRateProviderMock, GetFrameSeconds() ).WillByDefault( Return( 1 ) );
    }
@@ -39,6 +43,7 @@ public:
 protected:
    shared_ptr<mock_ConsoleBuffer> _consoleBufferMock;
    shared_ptr<mock_FrameRateProvider> _frameRateProviderMock;
+   shared_ptr<ConsoleSpriteDefs> _spriteDefs;
    shared_ptr<ConsoleRenderDefs> _renderDefs;
 
    shared_ptr<mock_ConsoleSprite> _particleSpriteMock;
