+++ MegaManLofiTests/PlayerThwipOutConsoleAnimationTests.cpp
@@ -4,6 +4,7 @@
 
 #include <MegaManLofi/PlayerThwipOutConsoleAnimation.h>
 #include <MegaManLofi/ConsoleRenderDefs.h>
+#include <MegaManLofi/ConsoleSpriteDefs.h>
 
 #include "mock_ConsoleBuffer.h"
 #include "mock_FrameRateProvider.h"
@@ -19,21 +20,23 @@ public:
    void SetUp() override
    {
       _renderDefs.reset( new ConsoleRenderDefs );
+      _spriteDefs.reset( new ConsoleSpriteDefs );
       _consoleBufferMock.reset( new NiceMock<mock_ConsoleBuffer> );
       _frameRateProviderMock.reset( new NiceMock<mock_FrameRateProvider> );
       _transitionSpriteMock.reset( new NiceMock<mock_ConsoleSprite> );
       _thwipSpriteMock.reset( new NiceMock<mock_ConsoleSprite> );
 
+      _spriteDefs->PlayerThwipOutTransitionSprite = _transitionSpriteMock;
+      _spriteDefs->PlayerThwipSprite = _thwipSpriteMock;
+
       _renderDefs->ArenaCharWidth = 1;
       _renderDefs->ArenaCharHeight = 1;
       _renderDefs->PlayerThwipVelocity = 1;
       _renderDefs->PlayerPostThwipDelaySeconds = 2;
+      _renderDefs->SpriteDefs = _spriteDefs;
 
       ON_CALL( *_frameRateProviderMock, GetFrameSeconds() ).WillByDefault( Return( 1 ) );
       ON_CALL( *_transitionSpriteMock, GetTotalTraversalSeconds() ).WillByDefault( Return( 10 ) );
-
-      _renderDefs->PlayerThwipOutTransitionSprite = _transitionSpriteMock;
-      _renderDefs->PlayerThwipSprite = _thwipSpriteMock;
    }
 
    void BuildAnimation()
@@ -43,6 +46,7 @@ public:
 
 protected:
    shared_ptr<mock_ConsoleBuffer> _consoleBufferMock;
+   shared_ptr<ConsoleSpriteDefs> _spriteDefs;
    shared_ptr<ConsoleRenderDefs> _renderDefs;
    shared_ptr<mock_FrameRateProvider> _frameRateProviderMock;
